style: reorder class members and extract SudokuFilterFlags for Konsist
Pure reordering (properties before functions, overrides before
non-overrides, companion object and nested classes last) to satisfy
CodingConventionsTest — no behavior change.

Also moves the TYPE_*/DIFFICULTY_*/SIZE_* filter-flag bit constants out
of GetAllSudokusUseCase's companion into domain/model/SudokuFilterFlags,
fixing a real data-depends-on-domain-use-case violation
(UserSettingsRepository referenced them only for defaults).
