+++ app/src/main/java/de/lemke/sudoku/domain/IsDailySudokuCompletedUseCase.kt
@@ -16,6 +16,9 @@
 
 package de.lemke.sudoku.domain
 
+import de.lemke.sudoku.domain.model.SudokuFilterFlags.DIFFICULTY_ALL
+import de.lemke.sudoku.domain.model.SudokuFilterFlags.SIZE_ALL
+import de.lemke.sudoku.domain.model.SudokuFilterFlags.TYPE_DAILY
 import java.time.LocalDate
 import javax.inject.Inject
 import kotlinx.coroutines.Dispatchers
@@ -27,7 +30,7 @@ class IsDailySudokuCompletedUseCase @Inject constructor(
     suspend operator fun invoke(date: LocalDate = LocalDate.now()): Boolean =
         withContext(Dispatchers.Default) {
             return@withContext getAllSudokus(
-                GetAllSudokusUseCase.TYPE_DAILY or GetAllSudokusUseCase.DIFFICULTY_ALL or GetAllSudokusUseCase.SIZE_ALL,
+                TYPE_DAILY or DIFFICULTY_ALL or SIZE_ALL,
             ).find { it.created.toLocalDate() == date && it.completed } != null
         }
 }
